requirements.txt
alembic==0.9.9
amqp==2.2.2
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
bcolz==1.2.0
better-exceptions==0.1.8
billiard==3.5.0.3
Bottleneck==1.2.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
cyordereddict==1.0.0
Cython==0.28.3
decorator==4.3.0
dill==0.2.5
Django==1.11.13
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
djangorestframework==3.8.2
empyrical==0.4.3
et-xmlfile==1.0.1
future==0.16.0
idna==2.6
intervaltree==2.1.0
ipython==6.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.3
jedi==0.12.0
Jinja2==2.10
jsonpickle==0.9.4
kiwisolver==1.0.1
kombu==4.1.0
line-profiler==2.1.2
Logbook==1.0.0
lru-dict==1.1.6
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
more-itertools==4.2.0
multipledispatch==0.5.0
mysqlclient==1.3.12
networkx==1.11
numexpr==2.6.5
numpy==1.14.3
openpyxl==2.5.1
pandas==0.18.1
pandas-datareader==0.5.0
parso==0.2.1
patsy==0.5.0
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.6.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
rqalpha==3.0.10
scipy==1.1.0
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
sortedcontainers==2.0.2
SQLAlchemy==1.2.7
statsmodels==0.9.0
tables==3.4.3
tabulate==0.8.2
toolz==0.9.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
XlsxWriter==1.0.5
